splash screen, dynamic txtview, new logo, notifications
I just created a new logo for our app, a better one, hopefully you guys
like it. And created a Splash screen, and figured out how to change the
textview content dynamically. So now out next and preview question
buttons work!! And also figured out how to show notification. It is
pretty simple which was Candy's part. So Candy!! just use
Toast.makeText() method and it should be inside the button events. I put
an example inside the signin button event. I'd appreciate if you add
notifications for all the textboxes and button events. Remember to
change the location of the notification, so it would show up right above
or bellow the textboxes. I'm gonna work on bettering the app design and
adding more layouts if it's necessary. And since nobody hasn't pushed
any code yet, so I didn't pull anything, just updating it!