changes
I changed the display of the drop down menu. I also took of the check
email and password notifications on the log in and I added a pop up if
your email and password dont match. I think for the drop down after you
select quiz and try to go back to account it wont go (or maybe my
computer was just being slow) can you test it on your phone Lala and let
me know what else needs to be fixed besides the colors.