new
i fixed that black pop ups and i added your code but when i tested it
out it always says that the email wasnt sent. i tried to fix it but
couldnt figure it out. i took out the edit profile from the spinner
since we have nothing for it.