timer added
Hey guys I added a count down timer. Its not the best but it was the
only thing I could find on the examples online. I put a 30 second time
and it goes to the next question when the time is up.I dont think i have
the most recent code so f you guys like it just change the
ClickerActivity.java and clicker.xml to the one on this project.